Add drawing Chain figures
Loop through the P list of a Chain object, draw lines. Easy.

I've made a slight change in how the Chain object works:
I deleted the color parameter of the constructor, and the user is
instead expected to write an optional color parameter as the last one,
as part of a one, integrated *args list. This is more intuitive, as it
matches the approach of all the other Figure derived classes.