Add basic Bezier curve drawing
Everything works perfectly with 3 points so far.
The implementation is still quite primitive, that's because I wasn't
expecting to reach my desired result so quickly and thus wanted to keep
things as simple as possible.

Fixed an error where the horizontal alignment settings would not apply
when drawing a specific state, as opposed to relying on the last state
(omitting the 'state' parameter).