Add Windows support
Changed the terminal size method to a module found on the internet,
because it's cross-platform.

Changed the color drawing to an implementation using termcolor and
Color_Console modules.

Fixed an error, when console window is too small, the message will print
maximum possible width correctly (halving it instead of doubling).